fix(app): move test-helpers out of __tests__ and fix flushMicrotasks
- Move test-helpers.ts from __tests__/ to test-utils/ so Jest does not
  discover it as a test suite with zero tests (CI failure)
- Replace setTimeout macrotask with Promise.resolve() for true
  microtask flush as documented
- Update all three test file imports to new path
